+++ manager/src/it/scala/io/hydrosphere/serving/manager/service/ModelManagementServiceITSpec.scala
@@ -2,19 +2,16 @@ package io.hydrosphere.serving.manager.service
 
 import io.hydrosphere.serving.contract.model_contract.ModelContract
 import io.hydrosphere.serving.manager.test.CommonIntegrationSpec
-import io.hydrosphere.serving.manager.model.api.description._
-import io.hydrosphere.serving.manager.model.api.ops.Implicits._
 import io.hydrosphere.serving.tensorflow.types.DataType
 import org.scalatest.BeforeAndAfterEach
 import io.hydrosphere.serving.contract.model_signature.ModelSignature
-import io.hydrosphere.serving.manager.model.api.{ContractBuilders, ModelType}
-
+import io.hydrosphere.serving.contract.utils.ContractBuilders
+import io.hydrosphere.serving.contract.utils.description._
+import io.hydrosphere.serving.manager.model.api.ModelType
+import io.hydrosphere.serving.contract.utils.ops.ModelContractOps._
 import scala.concurrent.Await
 import scala.concurrent.duration._
 
-/**
-  *
-  */
 class ModelManagementServiceITSpec extends CommonIntegrationSpec with BeforeAndAfterEach {
 
   describe("Model") {
